chore(network): rename block_header behaviour test file
